Smooth scrolling works.

+++ code/cpp/echomesh/color/Scroll.pyx
@@ -1,8 +1,8 @@
 cdef extern from "echomesh/color/Scroll.h" namespace "echomesh::color":
   FColorList scroll(FColorList, int dx, int dy, int xSize, bool wrap)
-  FColorList smoothScroll(FColorList, int dx, int dy, int xSize, bool wrap)
+  FColorList smoothScroll(FColorList, float dx, float dy, int xSize, bool wrap)
 
-def scroll_color_list(object fcl, int dx, int dy, int x_size,
+def scroll_color_list(object fcl, float dx, float dy, int x_size,
                       bool wrap=False, bool smooth=True):
   cdef ColorList source
   cdef ColorList result
@@ -12,5 +12,5 @@ def scroll_color_list(object fcl, int dx, int dy, int x_size,
   if smooth:
     result.thisptr.copy(smoothScroll(source.thisptr[0], dx, dy, x_size, wrap))
   else:
-    result.thisptr.copy(scroll(source.thisptr[0], dx, dy, x_size, wrap))
+    result.thisptr.copy(scroll(source.thisptr[0], int(dx), int(dy), x_size, wrap))
   return result
